# packages/shared/openapi.yaml
openapi: 3.1.0
info:
  title: DuSheng Panel API
  version: 0.1.0
  description: >
    REST contract for DuSheng Panel. The API manages users, device groups,
    nodes, tunnels, forwarding rules, protocol policies, install tokens,
    traffic samples, audit logs, and node-agent synchronization.
servers:
  - url: http://127.0.0.1:18888
    description: Local API server
  - url: https://panel.example.com
    description: Production reverse proxy
...
paths:
...
  /api/v1/tenants:
    get:
      tags: [Admin]
      operationId: listTenants
      summary: List tenants
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Tenant page.
          content: { application/json: { schema: { $ref: "#/components/schemas/TenantPage" } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Admin]
      operationId: createTenant
      summary: Create a tenant
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/TenantWrite" } } }
      responses:
        "201":
          description: Tenant created.
          content: { application/json: { schema: { $ref: "#/components/schemas/Tenant" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/tenants/{id}:
    put:
      tags: [Admin]
      operationId: updateTenant
      summary: Replace a tenant
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/TenantWrite" } } }
      responses:
        "200":
          description: Tenant updated.
          content: { application/json: { schema: { $ref: "#/components/schemas/Tenant" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Admin]
      operationId: deleteTenant
      summary: Delete an unreferenced tenant
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/tenants/{id}/traffic/reset:
    post:
      tags: [Admin]
      operationId: resetTenantTraffic
      summary: Reset the tenant billing period usage
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Reset tenant.
          content: { application/json: { schema: { $ref: "#/components/schemas/Tenant" } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/tenant-tunnel-grants:
    get:
      tags: [Admin]
      operationId: listTenantTunnelGrants
      summary: List tenant tunnel grants
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/TunnelId"
      responses:
        "200":
          description: Grant page.
          content: { application/json: { schema: { $ref: "#/components/schemas/TenantTunnelGrantPage" } } }
    post:
      tags: [Admin]
      operationId: createTenantTunnelGrant
      summary: Grant a tenant access to a tunnel and optional port range
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/TenantTunnelGrantWrite" } } }
      responses:
        "201":
          description: Grant created.
          content: { application/json: { schema: { $ref: "#/components/schemas/TenantTunnelGrant" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/tenant-tunnel-grants/{id}:
    put:
      tags: [Admin]
      operationId: updateTenantTunnelGrant
      summary: Replace a tenant tunnel grant
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/TenantTunnelGrantWrite" } } }
      responses:
        "200":
          description: Grant updated.
          content: { application/json: { schema: { $ref: "#/components/schemas/TenantTunnelGrant" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [Admin]
      operationId: deleteTenantTunnelGrant
      summary: Delete an unused tenant tunnel grant
      security: [{ BearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access or refresh JWT.
    AgentBearerAuth:
      type: http
      scheme: bearer
      description: Node token returned by agent registration.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Search:
      name: q
      in: query
      required: false
      schema:
        type: string
    Status:
      name: status
      in: query
      required: false
      schema:
        type: string
...
    TenantId:
      name: tenantId
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    TunnelId:
      name: tunnelId
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated principal is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NoContent:
      description: Resource deleted.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    BaseModel:
      type: object
      required: [id, createdAt, updatedAt]
      properties:
        id:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PageMeta:
      type: object
      required: [items, total, page, pageSize]
      properties:
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
...
    TenantPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/Tenant" } }
    TenantTunnelGrantPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/TenantTunnelGrant" } }
...
    Tenant:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - $ref: "#/components/schemas/TenantWrite"
        - type: object
          required: [usedBytes, quotaBlocked]
          properties:
            usedBytes: { type: integer, format: int64, minimum: 0 }
            periodStartedAt: { type: [string, "null"], format: date-time }
            nextResetAt: { type: [string, "null"], format: date-time }
            quotaBlocked: { type: boolean }
            quotaBlockedAt: { type: [string, "null"], format: date-time }
    TenantWrite:
      type: object
      required: [name, code]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        code: { type: string, pattern: "^[a-z0-9][a-z0-9_-]{1,59}$" }
        status: { type: string, enum: [active, suspended, disabled], default: active }
        trafficLimitBytes: { type: integer, format: int64, minimum: 0 }
        forwardLimit: { type: integer, minimum: 0 }
        userLimit: { type: integer, minimum: 0 }
        resetIntervalDays: { type: integer, minimum: 0, maximum: 3660 }
        expiresAt: { type: [string, "null"], format: date-time }
        notes: { type: string }
    TenantTunnelGrant:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - $ref: "#/components/schemas/TenantTunnelGrantWrite"
    TenantTunnelGrantWrite:
      type: object
      required: [tenantId, tunnelId]
      properties:
        tenantId: { type: integer, minimum: 1 }
        tunnelId: { type: integer, minimum: 1 }
        forwardLimit: { type: integer, minimum: 0 }
        portStart: { type: integer, minimum: 0, maximum: 65535 }
        portEnd: { type: integer, minimum: 0, maximum: 65535 }
...